+++ src/IO/ReadHelpers.cpp
@@ -1,4 +1,5 @@
 #include <Core/Defines.h>
+#include <Common/hex.h>
 #include <Common/PODArray.h>
 #include <Common/StringUtils/StringUtils.h>
 #include <Common/memcpySmall.h>
@@ -25,6 +26,59 @@ namespace ErrorCodes
     extern const int CANNOT_PARSE_DATETIME;
     extern const int CANNOT_PARSE_DATE;
     extern const int INCORRECT_DATA;
+}
+
+template <typename IteratorSrc, typename IteratorDst>
+void parseHex(IteratorSrc src, IteratorDst dst, const size_t num_bytes)
+{
+    size_t src_pos = 0;
+    size_t dst_pos = 0;
+    for (; dst_pos < num_bytes; ++dst_pos, src_pos += 2)
+        dst[dst_pos] = unhex2(reinterpret_cast<const char *>(&src[src_pos]));
+}
+
+void parseUUID(const UInt8 * src36, UInt8 * dst16)
+{
+    /// If string is not like UUID - implementation specific behaviour.
+
+    parseHex(&src36[0], &dst16[0], 4);
+    parseHex(&src36[9], &dst16[4], 2);
+    parseHex(&src36[14], &dst16[6], 2);
+    parseHex(&src36[19], &dst16[8], 2);
+    parseHex(&src36[24], &dst16[10], 6);
+}
+
+void parseUUIDWithoutSeparator(const UInt8 * src36, UInt8 * dst16)
+{
+    /// If string is not like UUID - implementation specific behaviour.
+
+    parseHex(&src36[0], &dst16[0], 16);
+}
+
+/** Function used when byte ordering is important when parsing uuid
+ *  ex: When we create an UUID type
+ */
+void parseUUID(const UInt8 * src36, std::reverse_iterator<UInt8 *> dst16)
+{
+    /// If string is not like UUID - implementation specific behaviour.
+
+    /// FIXME This code looks like trash.
+    parseHex(&src36[0], dst16 + 8, 4);
+    parseHex(&src36[9], dst16 + 12, 2);
+    parseHex(&src36[14], dst16 + 14, 2);
+    parseHex(&src36[19], dst16, 2);
+    parseHex(&src36[24], dst16 + 2, 6);
+}
+
+/** Function used when byte ordering is important when parsing uuid
+ *  ex: When we create an UUID type
+ */
+void parseUUIDWithoutSeparator(const UInt8 * src36, std::reverse_iterator<UInt8 *> dst16)
+{
+    /// If string is not like UUID - implementation specific behaviour.
+
+    parseHex(&src36[0], dst16 + 8, 8);
+    parseHex(&src36[16], dst16, 8);
 }
 
 UInt128 stringToUUID(const String & str)
